+++ src/DTOs/ping.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel, Field
 
 class Ping(BaseModel):
     status: str = Field(
-        default="not ok",
+        default="ok",
         description="Ping response of the server. Should be 'ok' in happy-case.",
-        examples=["not ok", "ok"],
+        examples=["ok", "not ok"],
     )